+++ src/utils.js
@@ -1,6 +1,5 @@
 const vscode = require('vscode');
-const strip = require('strip-comments');
-
+const { parseAbiFunctions, parsePrivateFunctions } = require("./parser");
 const snippetsJsonHover = require("./snippets/hover.json");
 const wordsSetHover = snippetsJsonHover['.source.ton-solidity'];
 
@@ -45,7 +44,7 @@ function getErrors(string) {
 }
 
 function highliteIt(functionSet) {
-    Object.keys(functionSet).map(function (k) {
+    Object.keys(functionSet).map((k) => {
         functionSet[k] = {
             prefix: functionSet[k].prefix,
             body: functionSet[k].body,
@@ -64,7 +63,7 @@ function getHoverItems(word, document) {
     let suggestion = null;
     let counter = 0;
     if (word.match(/AbiHeader|msgValue|pragma|(ton-)?solidity/)) {
-        let snippetPath = path.resolve(__dirname, "snippets/includes/" + word + ".md");
+        let snippetPath = path.resolve(__dirname, `snippets/includes/${word}.md`);
         suggestion = fs.readFileSync(snippetPath, "utf8");
         return suggestion;
     }
@@ -75,15 +74,18 @@ function getHoverItems(word, document) {
             if (value.prefix.length < prefixLength) {
                 continue;
             }
+            if (value.prefix.length == prefixLength) {
+                counter++;
+            }
             prefixLength = value.prefix.length;
             if (Array.isArray(value.description)) {
                 suggestion = value.description.join("\n")
             } else {
                 suggestion = value.description;
             }
-            counter++;
+            
             if (counter == 2) {
-                let snippetPath = path.resolve(__dirname, "snippets/includes/" + value.prefix.replace(/^\./, '') + ".md");
+                let snippetPath = path.resolve(__dirname, `snippets/includes/${value.prefix.replace(/^\./, '')}.md`);
                 suggestion = fs.readFileSync(snippetPath, "utf8");
                 return suggestion;
             }
@@ -109,68 +111,6 @@ function getSnippetItems(document) {
     }
 
     return completionItems;
-}
-
-function parseAbiFunctions(document) {
-    let abiPathData = path.parse(document.uri.fsPath);
-    let abiPath = path.resolve(__dirname, "abi/" + abiPathData.name + ".abi.json");
-    let abi;
-    try {
-        delete require.cache[abiPath]
-        abi = require(abiPath);
-    } catch (e) {
-        return {};
-    }
-    let completions = {};
-    for (const [, functionItem] of Object.entries(abi.functions)) {
-        if (functionItem.name == 'constructor') {
-            continue;
-        }
-        //public variable
-        if (functionItem.outputs.length == 1 && functionItem.outputs[0].name == functionItem.name) {
-            continue;
-        }
-        let key = 'function ' + functionItem.name;
-        let prefix = functionItem.name;
-        let paramsBody = [];
-        let paramsInputDescription = [];
-        //input
-        for (const [index, inputItem] of Object.entries(functionItem.inputs)) {
-            paramsBody.push('${' + (Number(index) + 1) + ':' + inputItem.type + ' ' + inputItem.name + '}');
-            paramsInputDescription.push(inputItem.type + ':' + inputItem.name);
-        }
-        let body = functionItem.name + '(' + paramsBody.join(', ') + ')';
-        let description = 'function ' + functionItem.name + '(' + paramsInputDescription.join(', ') + ')';
-        //output
-        let paramsOutputDescription = [];
-        for (const [, outputItem] of Object.entries(functionItem.outputs)) {
-            paramsOutputDescription.push(outputItem.type);
-        }
-        description += paramsOutputDescription.length > 0 ? ': ' + paramsOutputDescription.join(', ') : '';
-        completions[key] = { prefix, body, description };
-    }
-
-    return completions;
-}
-
-function parsePrivateFunctions(document) {
-    let code = document.getText();
-    code = strip(code);
-    let privateFunctions = {};
-    let matches = [...code.matchAll(/((function\s+([a-z_A-Z0-9]+))\((.*)\)(.*private.*)){/gm)];
-    for (let item of matches) {
-        let params = item[4].split(",");
-        let fparams = params.map((value, index) => {
-            return '${' + Number(index + 1) + ':' + value + '}';
-        })
-        let importParams = fparams.join(", ");
-        privateFunctions[item[2]] = {
-            prefix: item[3],
-            body: item[2] + '(' + importParams + ')',
-            description: item[1]
-        }
-    }
-    return privateFunctions;
 }
 
 function getSnippetType(body) {
